backend/requirements.txt
agno
openai
fastapi
uvicorn
python-dotenv
pydantic
rich>=13.0.0

